requirements.txt
fastapi==0.92
uvicorn==0.20.0
SQLAlchemy==2.0.4
psycopg2==2.9.5
python_dotenv==0.21.1
fastapi_jwt_auth[asymmetric]==0.5.0
email_validator==1.3.1
passlib==1.7.4
alembic==1.9.4

# Testing packages
pytest==7.2.1
factory-boy==3.2.1
httpx==0.23.3
pytest-postgresql==4.1.1